Rename optimal_scores_generator.py to score_utils.py
